script/sync_prod_to_gcp/requirements.txt
arxiv-base==0.17.4.post2
attrs==22.1.0
backports-datetime-fromisoformat==1.0.0
bleach==3.3.0
boto3==1.23.10
botocore==1.26.10
cachetools==4.2.4
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
google-api-core==2.10.2
google-auth==2.14.1
google-cloud==0.34.0
google-cloud-audit-log==0.2.2
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
google-cloud-pubsub
googleapis-common-protos==1.56.3
idna==3.4
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.0.0
logging-json==0.4.0
MarkupSafe==2.0.1
mypy==0.971
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.11.2
proto-plus==1.22.1
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.27.1
rsa==4.9
s3transfer==0.5.2
six==1.16.0
tomli==1.2.3
typed-ast==1.5.4
typing-extensions==4.1.1
urllib3==1.26.12
uWSGI==2.0.21
webencodings==0.5.1
Werkzeug==2.0.3
zipp==3.6.0
lmdb==1.4.1

